+++ CHANGES.rst
@@ -3,6 +3,13 @@ Change log
 
 .. note:: **Upgrade notes**: after upgrading, run the ``arouteserver setup-templates`` command to sync the local templates with those distributed with the new version. More details on the `Upgrading <https://arouteserver.readthedocs.io/en/latest/INSTALLATION.html#upgrading>`__ section of the documentation.
 
+next release
+------------
+
+- Improvement: ``configure`` command.
+  
+  It can be used to quickly generate a route server policy definition file (``general.yml``) on the basis of best practices and suggestions.
+
 v0.14.1
 -------